refactor(app): update routing paths in App component
