+++ src/layouts/Sidebar.jsx
@@ -18,7 +18,7 @@ const Sidebar = () => {
         { name: "Crear Posición", icon: Icons.sidebar.positionCreate, path: "/position" },
         { name: "Subir CVs", icon: Icons.sidebar.uploadCv, path: "/uploadcv" },
         { name: "Crear Vacante", icon: Icons.sidebar.vacant, path: "/create-vacant" },
-        { name: "Historial de CVs", icon: Icons.sidebar.eyeHistory, path: "/cv-history" },
+        { name: "Historial de CVs", icon: Icons.sidebar.history, path: "/cv-history" },
         { name: "Historial de anuncios", icon: Icons.sidebar.historyVacant, path: "/ads-history" },
         { name: "Historial de Posiciones", icon: Icons.sidebar.historyPosition, path: "/position-history" },
         { name: "Resultados", icon: Icons.sidebar.trophy, path: "/results" },
